requirements.txt
Flask
openai
python-dotenv
rapidfuzz
